LZMR-47 input request mappings for organization controller
